CI: don't unilateraly write to /tmp/vol
This worked both for local & CI runs since all runs were started through
dockertest.sh.
Now that we are migrating to k8s runners, we can't rely on a docker volume, and
can't assume that CI_PROJECT_DIR will be present as dockertest.sh can still be
used for local testing.
This PR adds a TESTING_DIR environment variable that points the scripts at the
proper location, and defaults to CI_PROJECT_DIR when it's defined, or /tmp/vol
when ran from dockertest.sh