fix Red Hat spelling
It is spelled as 2 words in the files we inspect to determine the
distribution we're running on.
Fixing the search pattern changes the value assigned to DISTRIBUTION so
all existing checks have been updated.
OS value was also update for consistency, so we don't have to wonder if
we need to spell Red Hat a one or two words