[SINT-3627] Support TESTING_KEYS_URL in E2E tests (#355)
https://github.com/DataDog/agent-linux-install-script/pull/353 changed E2E tests to use testing packages from the upstream pipeline (if there is one).

In datadog-agent, we are soon going to use a dedicated testing location for signing keys, which needs to be passed to any test that use testing packages built by the datadog-agent pipeline. Therefore, the TESTING_KEYS_URL variable needs to be passed to the install script used in the E2E test.